Separated all the tests and made it a package
