+++ .idea/runConfigurations/Unit_Test__Command___CSV.xml
@@ -1,5 +1,5 @@
 <component name="ProjectRunConfigurationManager">
-  <configuration default="false" name="Tests: Unit" type="tests" factoryName="Unittests">
+  <configuration default="false" name="Unit Test (Command): CSV" type="tests" factoryName="Unittests">
     <module name="Table2XSV" />
     <option name="INTERPRETER_OPTIONS" value="" />
     <option name="PARENT_ENVS" value="true" />
@@ -10,7 +10,7 @@
     <option name="ADD_SOURCE_ROOTS" value="true" />
     <option name="_new_pattern" value="&quot;&quot;" />
     <option name="_new_additionalArguments" value="&quot;&quot;" />
-    <option name="_new_target" value="&quot;$PROJECT_DIR$/tests/tests.py&quot;" />
+    <option name="_new_target" value="&quot;$PROJECT_DIR$/tests/unit/test_csv_command.py&quot;" />
     <option name="_new_targetType" value="&quot;PATH&quot;" />
     <method v="2" />
   </configuration>
