+++ .idea/workspace.xml
@@ -1,9 +1,12 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project version="4">
   <component name="ChangeListManager">
-    <list default="true" id="ad6b76c0-59b7-4bf4-9448-d2a6b18bda1a" name="Changes" comment="Shows that project supports Python 3.10">
+    <list default="true" id="ad6b76c0-59b7-4bf4-9448-d2a6b18bda1a" name="Changes" comment="Add Python 3.10 Classifier">
+      <change afterPath="$PROJECT_DIR$/tests/unit/__init__.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/.idea/runConfigurations/Tests__Unit.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/runConfigurations/Unit_Test__Command___CSV.xml" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/.idea/workspace.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/workspace.xml" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/setup.cfg" beforeDir="false" afterPath="$PROJECT_DIR$/setup.cfg" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/tests/tests.py" beforeDir="false" />
     </list>
     <option name="SHOW_DIALOG" value="false" />
     <option name="HIGHLIGHT_CONFLICTS" value="true" />
@@ -20,7 +23,7 @@
   <component name="Git.Settings">
     <option name="RECENT_BRANCH_BY_REPOSITORY">
       <map>
-        <entry key="$PROJECT_DIR$" value="main" />
+        <entry key="$PROJECT_DIR$" value="development" />
       </map>
     </option>
     <option name="RECENT_GIT_ROOT_PATH" value="$PROJECT_DIR$" />
@@ -43,7 +46,7 @@
     <property name="RunOnceActivity.OpenProjectViewOnStart" value="true" />
     <property name="RunOnceActivity.ShowReadmeOnStart" value="true" />
     <property name="SHARE_PROJECT_CONFIGURATION_FILES" value="true" />
-    <property name="last_opened_file_path" value="$PROJECT_DIR$" />
+    <property name="last_opened_file_path" value="$PROJECT_DIR$/tests/unit/test_csv_command.py" />
     <property name="settings.editor.selected.configurable" value="com.jetbrains.python.configuration.PyActiveSdkModuleConfigurable" />
   </component>
   <component name="RecentsManager">
@@ -55,9 +58,14 @@
       <recent name="C:\Users\amogh.madan\PycharmProjects\Table2XSV\table2xsv\management" />
     </key>
   </component>
-  <component name="RunManager" selected="tox.TOX (Checker): Black (Format)">
+  <component name="RunManager" selected="tox.TOX (Build): Package">
     <list>
-      <item itemvalue="Python tests.Tests: Unit" />
+      <item itemvalue="Python tests.Unit Test (Command): Base" />
+      <item itemvalue="Python tests.Unit Test (Command): CSV" />
+      <item itemvalue="Python tests.Unit Test (Command): Excel" />
+      <item itemvalue="Python tests.Unit Test (Command): MySQL" />
+      <item itemvalue="Python tests.Unit Test (Command): Neo4j" />
+      <item itemvalue="Python tests.Unit Test (Command): SQLite" />
       <item itemvalue="tox.TOX (Build): Executable [All]" />
       <item itemvalue="tox.TOX (Build): Package" />
       <item itemvalue="tox.TOX (Checker): Bandit" />
